Updated to include reading carpark features
added read address, parking system type, parking available time, free parking, carpark rate etc. However, the reason some of them are commented out is due to a problem with the data column name, I'll have to clean it to not start with a number, or else I cannot call the feature of that document.